Handle incorrectly named/missing columns, add tests
Improved the code to handle incorrect or missing columns - for cases
where users try to upload a file with columns that are not defined on
in the annotations.
Added test cases for the various exceptions and their messages.
Finally, instances of ZeroCellException are just bubbled up instead of
being wrapped in another ZCE