Add annotation processor for ZeroCell
Implemented an annotation processor that generates
a ZeroCellReader implementation as a source file -
this will enable better tuning of the implementation
and also provide visibility for auditing.

Might be a breaking change for existing ZeroCell users
if the use was more than just the annotations and the Reader
e.g. Converter which has a modified API now.