Modify ZeroCellReader API, bump version to 0.2.3-SNAPSHOT
Modified the ZeroCellReader API to take in a file and
sheet as arguments so that the generated readers have
a simpler API.
The reader should be reusable across files and shouldn't
require re-allocating a new one.
Decided to just bump the version to 0.2.3 since the API
doesn't break major things and there no real users of
the 0.2.x version of the library at the moment.