Remove "defaults" in the default Converters
The default Converters had this concept of defaults when the
parsing of the value from the cell failed - in this change
we remove defaults in favour of returning nulls. Perhaps
it would be better to throw an exception but for null this
should suffice.
Users should handle invalid values themselves