Add Fallback strategy and refactor converters (#21)
* Add converter classes that take in a fallback strategy

Moved the logic of the default converters outside lambdas to separate
classes and added a DefaultConverter superclass. Each DefaultConverter
requires a fallbackStrategy which tells the library how do deal with
null values and other invalid values.

* Update exception message and log failure to convert value

* Remove logging statement

* Add test for boolean converter

* Update version to 0.5.0-SNAPSHOT

* Add constructor that takes in FallbackStrategy

* Add FallbackStrategy.IGNORE to allow users to ignore invalid values.

Added a fallback strategy to allow user to ignore invalid values
and continue processing the file. Added documentation for each strategy.
Added the FallbackStrategy to the default convertors

* Move logic for writing to instance field into own class

Moved the logic for writing to the fields into a separate class
to make it more testable. Also added a test that uses this new
class and the FallbackStrategy.IGNORE strategy

* Update docblock

* added trim and uppercase method to be return when getting the column name

* Fix converters and tests

* Add converter classes that take in a fallback strategy

Moved the logic of the default converters outside lambdas to separate
classes and added a DefaultConverter superclass. Each DefaultConverter
requires a fallbackStrategy which tells the library how do deal with
null values and other invalid values.

* Update exception message and log failure to convert value

* Remove logging statement

* Add test for boolean converter

* Update version to 0.5.0-SNAPSHOT

* Add constructor that takes in FallbackStrategy

* Add FallbackStrategy.IGNORE to allow users to ignore invalid values.

Added a fallback strategy to allow user to ignore invalid values
and continue processing the file. Added documentation for each strategy.
Added the FallbackStrategy to the default convertors

* Move logic for writing to instance field into own class

Moved the logic for writing to the fields into a separate class
to make it more testable. Also added a test that uses this new
class and the FallbackStrategy.IGNORE strategy

* Update docblock

* Repalce excelant dependency with ooxml

* Fix tests to use convertToType method